+++ Group_project/Group_project/Debug/Group_project.log
@@ -1,4 +1,12 @@
 ﻿  main.cpp
-  Threads.cpp
-  Generating Code...
-  Group_project.vcxproj -> D:\GitHub\meme_party\Group_project\Debug\Group_project.exe
+D:\GitHub\meme_party\Group_project\Group_project\IteratorCSV.h(18,9): error C3861: 'm_it_': identifier not found
+D:\GitHub\meme_party\Group_project\Group_project\IteratorCSV.h(33): message : see reference to class template instantiation 'Iterator<T,U>' being compiled
+D:\GitHub\meme_party\Group_project\Group_project\IteratorCSV.h(21,9): error C3861: 'm_it_': identifier not found
+D:\GitHub\meme_party\Group_project\Group_project\IteratorCSV.h(23,19): error C2065: 'm_it_': undeclared identifier
+D:\GitHub\meme_party\Group_project\Group_project\IteratorCSV.h(23,19): error C3861: 'm_it_': identifier not found
+D:\GitHub\meme_party\Group_project\Group_project\IteratorCSV.h(26,17): error C3861: 'm_it_': identifier not found
+D:\GitHub\meme_party\Group_project\Group_project\IteratorCSV.h(29,28): error C3861: 'm_it_': identifier not found
+D:\GitHub\meme_party\Group_project\Group_project\IteratorCSV.h(64,19): warning C4018: '<': signed/unsigned mismatch
+D:\GitHub\meme_party\Group_project\Group_project\IteratorCSV.h(72,125): error C2440: '<function-style-cast>': cannot convert from 'initializer list' to 'std::pair<std::string,Matrix<double> &>'
+D:\GitHub\meme_party\Group_project\Group_project\IteratorCSV.h(72,125): message : No constructor could take the source type, or constructor overload resolution was ambiguous
+D:\GitHub\meme_party\Group_project\Group_project\IteratorCSV.h(72,29): error C2064: term does not evaluate to a function taking 2 arguments
