+++ Group_project/Group_project/main.cpp
@@ -5,13 +5,20 @@
 #include <map>
 #include <iostream>
 #include "CSVReader.h"
+#include "IteratorCSV.h"
 
 int main()
 {
-	map<string, algorithm*> task;
+	/*map<string, algorithm*> task;
 	task["files/test1.csv"] = new RKintegrator<RKmethod_5_4_7fs, Osc_Van_der_Pol>;
 	task["files/test2.csv"] = new RKintegrator<VernerRK_5_6_8, LVOde>;
-	paralel_alg<1>(task);
+	paralel_alg<1>(task);*/
+	Matrix<double> mat(3, 3, 0);
+	mat[0][0] = 1;
+	mat[1][1] = 2;
+	mat[2][2] = 3;
+	Matrix<double> a = mat.slice(2, 2);
+	cout << a << endl;
 	system("pause");
 	return 0;
 }
